+++ modules/shared_vpc/main.tf
@@ -56,6 +56,7 @@ module "project-factory" {
   disable_dependent_services        = var.disable_dependent_services
   python_interpreter_path           = var.python_interpreter_path
   use_tf_google_credentials_env_var = var.use_tf_google_credentials_env_var
+  skip_gcloud_download              = var.skip_gcloud_download
 }
 
 /******************************************